// src/features/space/space.type.ts
export type ITargetType =
  | 'GENERAL_LEARNING'
  | 'TEST_PREPARATION'
  | 'BUSINESS'
  | 'ACADEMIC'
  | 'CONVERSATION'
  | 'TRAVEL'
  | 'PROFESSIONAL'
  | 'ORTHER'

export type ILanguageType =
  | 'ENGLISH'
  | 'JAPANESE'
  | 'KOREAN'
  | 'CHINESE'
  | 'VIETNAMESE'
  | 'FRENCH'
  | 'GERMAN'
  | 'SPANISH'

export enum LanguageCode {
  ENGLISH = 'ENGLISH',
  VIETNAMESE = 'VIETNAMESE'
}

export enum LevelCode {
  BEGINNER = 'BEGINNER',
  ELEMENTARY = 'ELEMENTARY',
  INTERMEDIATE = 'INTERMEDIATE',
  UPPER_INTERMEDIATE = 'UPPER_INTERMEDIATE',
  ADVANCED = 'ADVANCED',
  MASTER = 'MASTER'
}

export enum TargetCode {
  COMMUNICATION = 'COMMUNICATION',
  IELTS = 'IELTS',
  TOEIC = 'TOEIC',
  OTHER = 'OTHER'
}

export enum TopicCode {
  BUSINESS = 'BUSINESS',
  ACADEMIC = 'ACADEMIC',
  TRAVEL = 'TRAVEL',
  DAILY_LIFE = 'DAILY_LIFE',
  TECHNOLOGY = 'TECHNOLOGY',
  OTHER = 'OTHER'
}